meeting-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
httpx==0.28.1
redis==5.2.1
celery==5.4.0
pytest==8.3.4
